backend/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
google-api-python-client
google-auth
python-dotenv
httpx
pytest
pytest-asyncio
google-genai>=0.8.0
Pillow>=10.0.0
python-multipart>=0.0.9
